pyproject.toml
[tool.poetry]
name = "gist-storage"
description = "A versatile tool designed to store and share data via gists, ideal for synchronizing information across devices. Since gists are publicly available, it includes an option for encryption to ensure security."
version = "0.1.1"
license = "MIT"

authors = []

readme = "README.md"

repository = "https://github.com/psychonaute/gist-storage"

keywords = []

classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Operating System :: OS Independent",
  "Topic :: Software Development :: Libraries :: Python Modules",
]

[tool.poetry.dependencies]
python = "^3.11"
pygithub = "^2.1.1"
cryptography = { version = "^42.0.2", optional = true }
python-dotenv = "^1.0.1"

[tool.poetry.extras]
encryption = ["cryptography", "python-dotenv"]

[tool.poetry.group.test.dependencies]
nitpick = "^0.32"

safety = "^2.3"

pytest = "^7.1"
pytest-cov = "^4.0"
pytest-randomly = "^3.12"


[tool.poetry.group.dev.dependencies]
types-requests = "^2.31.0.20240125"
pytest-mock = "^3.12.0"

[build-system]
requires = ["poetry-core>=1.1.13"]
build-backend = "poetry.core.masonry.api"

# add third party modules twhom stud file can't be found
# cf: Python - typing.xmind
[[tool.mypy.overrides]]
module = []
ignore_missing_imports = true
